server/requirements.txt
# Server Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
